Added list, addtime, expired, +fixes
v0.6
* Fixed the help to only show commands you are allowed to use
* Added command "/plotme list" which displays plots that you are allowed
to build on along with the expiration date and helpers.
(plotme.user.list)
* Added command "/plotme list [player]" which displays plots that you
are allowed to build on along with the expiration date and helpers.
(plotme.admin.list)
* Added command "/plotme expired [page]" which displays expired plots
(plotme.admin.expired)
* Added command "/plotme addtime" which resets the expiration date
depending on the DaysToExpiration configuration. (plotme.admin.addtime)
* Plots which get modified by the plot owner or by one of the helpers
get it's expiration date reset. (Listening to block place, break and
interact)
* Added new parameter in the config.yml for the DaysToExpiration.
Default is 7. Set to 0 for never expire.
* Fixed config.yml to input default values for new parameters when they
were previously not present (Example: WorldHeight)
* Changed the date format for Expiration date in the command "/plotme
info" to YYYY-MM-DD